refactor: rename bits and bobs
IEntityDbContext, EntityDbContextBase

does not explicitly depend on involving snapshots
