+++ src/EntityDb.EntityFramework/DbContexts/IEntityDbContext.cs
@@ -1,0 +1,19 @@
+﻿using EntityDb.EntityFramework.Sessions;
+using Microsoft.EntityFrameworkCore;
+
+namespace EntityDb.EntityFramework.DbContexts;
+
+/// <summary>
+///     A type of a <see cref="DbContext"/> that can be used for EntityDb purposes.
+/// </summary>
+/// <typeparam name="TDbContext">The type of the <see cref="DbContext"/></typeparam>
+public interface IEntityDbContext<TDbContext>
+    where TDbContext : DbContext, IEntityDbContext<TDbContext>
+{
+    /// <summary>
+    ///     Returns a new <typeparamref name="TDbContext"/> that will be configured using <paramref name="entityFrameworkSnapshotSessionOptions"/>.
+    /// </summary>
+    /// <param name="entityFrameworkSnapshotSessionOptions">The options for the database</param>
+    /// <returns>A new <typeparamref name="TDbContext"/> that will be configured using <paramref name="entityFrameworkSnapshotSessionOptions"/>.</returns>
+    static abstract Task<TDbContext> ConstructAsync(EntityFrameworkSnapshotSessionOptions entityFrameworkSnapshotSessionOptions);
+}
